refactor:[BE]Collection 디렉토리 변경
