+++ backend/src/main/java/com/example/webduck/collection/infrastructure/repository/CollectionRepositoryImpl.java
@@ -1,6 +1,7 @@
-package com.example.webduck.collection.infrastructure;
+package com.example.webduck.collection.infrastructure.repository;
 
 import com.example.webduck.collection.domain.Collection;
+import com.example.webduck.collection.infrastructure.entity.CollectionEntity;
 import com.example.webduck.collection.service.port.CollectionRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
