+++ backend/src/main/java/com/example/webduck/collection/service/port/CollectionWebtoonsRepository.java
@@ -1,6 +1,6 @@
 package com.example.webduck.collection.service.port;
 
-import com.example.webduck.collection.infrastructure.CollectionWebtoons;
+import com.example.webduck.collection.infrastructure.entity.CollectionWebtoons;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CollectionWebtoonsRepository extends JpaRepository<CollectionWebtoons,Long> {
